+++ src/app/(authorized)/layout.tsx
@@ -1,0 +1,9 @@
+import MainLayout from '@ui/layout/MainLayout';
+
+export default function AuthorizedLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <MainLayout>{children}</MainLayout>;
+}
