+++ src/app/(sentry)/sentry-example-page/page.tsx
