+++ src/app/(unauthorized)/layout.tsx
@@ -1,0 +1,7 @@
+export default function UnAuthorizedLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return <>{children}</>;
+}
