+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-import { authOptions } from '@/app/lib/authOptions';
+import { authOptions } from '@lib/authOptions';
 import NextAuth from 'next-auth/next';
 
 const handler = NextAuth(authOptions);
