+++ src/provider/QueryProvider.tsx
@@ -1,7 +1,7 @@
 'use client';
 
+import { queryClient } from '@lib/api/queryClient';
 import { QueryClientProvider } from '@tanstack/react-query';
-import { queryClient } from '../api/queryClient';
 
 export default function QueryProvider({
   children,
