+++ src/store/popup.store.ts
